Add mocks

+++ rest-api/src/services/meetings/meetings.service.ts
@@ -5,12 +5,18 @@ import MeetingDto from 'src/dto/Meeting';
 import { Meeting as MeetingEntity } from '../../entity/Meeting';
 import { meeting as meetingEntityToDomainMapper } from '../../util/mapper/entityToDomain';
 import { meeting as meetingDomainToEntityMapper } from '../../util/mapper/domainToEntity';
+import meetingMocks from './meetingMocks';
 
 @injectable()
 class MeetingsService implements IMeetingsService {
-  private meetingRepository = getRepository(MeetingEntity);
+  private meetingRepository = process.env.MOCKS ? undefined : getRepository(MeetingEntity);
 
   async getMeetings(): Promise<MeetingDto[]> {
+    if(process.env.MOCKS) {
+      return [meetingMocks.meetingMock(), meetingMocks.meetingMock()];
+    }
+
+    // @ts-ignore
     const meetingEntities = await this.meetingRepository.find({
       join: {
         alias: 'meeting',
@@ -32,6 +38,10 @@ class MeetingsService implements IMeetingsService {
   }
 
   async getMeetingById(id: string): Promise<MeetingDto | undefined> {
+    if(process.env.MOCKS) {
+      return meetingMocks.meetingMock();
+    }
+    // @ts-ignore
     const meetingEntity = await this.meetingRepository.findOne(id, {
       join: {
         alias: 'meeting',
@@ -47,27 +57,42 @@ class MeetingsService implements IMeetingsService {
   }
 
   async createMeeting(createMeetingRequest: MeetingDto): Promise<MeetingDto> {
+    if(process.env.MOCKS) {
+      return meetingMocks.meetingMock();
+    }
     const meetingEntity = meetingDomainToEntityMapper.map(createMeetingRequest);
+    // @ts-ignore
     const entityResponse = await this.meetingRepository.save(meetingEntity);
     return meetingEntityToDomainMapper.map(entityResponse);
   }
 
   async updateMeeting(updateMeetingRequest: MeetingDto): Promise<MeetingDto> {
+    if(process.env.MOCKS) {
+      return meetingMocks.meetingMock();
+    }
     const meetingEntityToSave = meetingDomainToEntityMapper.map(updateMeetingRequest);
+    // @ts-ignore
     const meetingEntity = await this.meetingRepository.findOne(meetingEntityToSave.id);
     let updatedMeetingEntity: MeetingEntity;
     if (meetingEntity && meetingEntity.id) {
+      // @ts-ignore
       await this.meetingRepository.update(meetingEntity.id, meetingEntityToSave); // bug here where address is not saved, cascade on update isn't working
       updatedMeetingEntity = meetingEntityToSave;
     } else {
+      // @ts-ignore
       updatedMeetingEntity = await this.meetingRepository.save(meetingEntityToSave);
     }
     return meetingEntityToDomainMapper.map(updatedMeetingEntity);
   }
 
   async deleteMeeting(id: string): Promise<MeetingDto | undefined> {
+    if(process.env.MOCKS) {
+      return meetingMocks.meetingMock();
+    }
+    // @ts-ignore
     const meetingEntity = await this.meetingRepository.findOne(id);
     if (meetingEntity) {
+      // @ts-ignore
       await this.meetingRepository.remove(meetingEntity);
       return meetingEntityToDomainMapper.map(meetingEntity);
     }
